sm_kyc_py_latest/requirements.txt
mysql-connector
pandas~=1.1.1
google-cloud-vision==0.35.2
pillow~=7.2.0
selenium~=3.141.0
flask~=1.1.2
boto3~=1.14.27
pika~=1.1.0
SQLAlchemy==1.3.18
urllib3==1.25.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
#PyAutoGUI==0.9.50
docutils==0.16
mouse==0.7.1
google-api-core==1.22.0
google-auth==1.19.2
python-xlib==0.28
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
jmespath==0.10.0
##MouseInfo==0.1.3
numpy==1.19.1
opencv-python==4.2.0.34
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
pyvirtualdisplay
azcaptchaapi==0.3.1
pdfplumber==0.5.22


botocore~=1.17.27